feat(fvS*): Add scheme entry for laplacian(De, pD)...
According to the changs made to the solver, I add an entry in the
fvScheme file for the laplacian term added to the pDEqn. I also change
the solver for pDEqn in fvSoltion file from DIC to DILC. The reason was
OpenFOAM raised an error that it cannot solve pDEqn with DIC and
provided some options. I chose DILC, for no reason!